Add ProtoDecodingContainer to centralize common decoding container functionality
